修改发布作业请求类的时间定义范围

+++ src/main/java/com/hebutgo/ework/entity/request/AnnounceDemandRequest.java
@@ -32,10 +32,10 @@ public class AnnounceDemandRequest {
     private Timestamp endTime;
 
     @ApiModelProperty(value = "开始（收作业）时间（整数格式）")
-    private Integer startTimeMills;
+    private Double startTimeMills;
 
     @ApiModelProperty(value = "结束（收作业）时间（整数格式）")
-    private Integer endTimeMills;
+    private Double endTimeMills;
 
     @ApiModelProperty(value = "发放小组id")
     private Integer groupId;
